Ext merge

+++ src/solvers/amg/Aggregation.hpp
@@ -213,7 +213,7 @@ std::vector<size_t> argsort( const std::vector<T> &array )
 {
     std::vector<size_t> indices( array.size() );
     std::iota( indices.begin(), indices.end(), 0 );
-    std::sort( indices.begin(), indices.end(), [&array]( int left, int right ) -> bool {
+    std::sort( indices.begin(), indices.end(), [&array]( size_t left, size_t right ) -> bool {
         // sort indices according to corresponding array element
         return array[left] < array[right];
     } );
